Split the Steam relay, occupancy state and net upgrades
  SteamRelayTransport (1180) -> transport, governor, connections, io, lifecycle
  ResidentialOccupancySyncSystem (797) -> state, cycle
  NetUpgradeSyncSystem (765) -> system, capture, apply; and into its own folder

SteamRelayTransport becomes partial to allow the split - the only change to it
beyond moving code.

Pure code movement, verified member by member against the previous revision.
